Simplifying the code around route controller handler
