--- src/controllers/grpc_route.rs
@@ -1,293 +1,0 @@
-use std::sync::Arc;
-
-use async_trait::async_trait;
-use futures::{future::BoxFuture, FutureExt, StreamExt};
-use gateway_api::{
-    common_types::{ParentRouteStatus, RouteRef, RouteStatus},
-    gateways::Gateway,
-    grpcroutes::{self, GRPCRoute},
-};
-use kube::{
-    runtime::{controller::Action, watcher::Config, Controller},
-    Api, Client, Resource,
-};
-use tokio::sync::mpsc::{self};
-use tracing::{debug, warn, Span};
-use typed_builder::TypedBuilder;
-use uuid::Uuid;
-
-use super::{
-    handlers::ResourceHandler,
-    utils::{ResourceCheckerArgs, ResourceState, RouteListenerMatcher},
-    ControllerError, RECONCILE_LONG_WAIT,
-};
-use crate::{
-    common::{self, ReferenceValidateRequest, RequestContext, ResourceKey, Route, RouteRefKey, VerifiyItems},
-    controllers::routes_common,
-    services::patchers::{DeleteContext, Operation},
-    state::State,
-};
-
-type Result<T, E = ControllerError> = std::result::Result<T, E>;
-
-#[derive(Clone, TypedBuilder)]
-pub struct GRPCRouteControllerContext {
-    controller_name: String,
-    client: Client,
-    state: State,
-    grpc_route_patcher: mpsc::Sender<Operation<GRPCRoute>>,
-    validate_references_channel_sender: mpsc::Sender<ReferenceValidateRequest>,
-}
-
-#[derive(TypedBuilder)]
-pub struct GRPCRouteController {
-    ctx: Arc<GRPCRouteControllerContext>,
-}
-
-impl GRPCRouteController {
-    pub fn get_controller(&self) -> BoxFuture<()> {
-        let client = self.ctx.client.clone();
-        let context = &self.ctx;
-
-        Controller::new(Api::all(client), Config::default())
-            .run(Self::reconcile_grpc_route, Self::error_policy, Arc::clone(context))
-            .for_each(|_| futures::future::ready(()))
-            .boxed()
-    }
-
-    #[allow(clippy::needless_pass_by_value)]
-    fn error_policy<T>(_object: Arc<T>, _err: &ControllerError, _ctx: Arc<GRPCRouteControllerContext>) -> Action {
-        Action::requeue(RECONCILE_LONG_WAIT)
-    }
-
-    async fn reconcile_grpc_route(resource: Arc<grpcroutes::GRPCRoute>, ctx: Arc<GRPCRouteControllerContext>) -> Result<Action> {
-        let controller_name = ctx.controller_name.clone();
-        let grpc_route_patcher = ctx.grpc_route_patcher.clone();
-
-        let Some(maybe_id) = resource.metadata.uid.clone() else {
-            return Err(ControllerError::InvalidPayload("Uid must be present".to_owned()));
-        };
-
-        let Ok(_) = Uuid::parse_str(&maybe_id) else {
-            return Err(ControllerError::InvalidPayload("Uid in wrong format".to_owned()));
-        };
-        let version = resource.meta().resource_version.clone();
-        let resource_key = ResourceKey::from(&*resource);
-
-        let state = &ctx.state;
-
-        let maybe_stored_route = state.get_grpc_route_by_id(&resource_key).expect("We expect the lock to work");
-
-        let _ = Route::try_from(&*resource)?;
-
-        let handler = GRPCRouteHandler::builder()
-            .state(ctx.state.clone())
-            .resource_key(resource_key)
-            .controller_name(controller_name)
-            .resource(resource)
-            .grpc_route_patcher(grpc_route_patcher)
-            .validate_references_channel_sender(ctx.validate_references_channel_sender.clone())
-            .version(version)
-            .build();
-
-        handler.process(maybe_stored_route, Self::check_spec, Self::check_status).await
-    }
-
-    fn check_spec(args: ResourceCheckerArgs<GRPCRoute>) -> ResourceState {
-        let (resource, stored_resource) = args;
-        if resource.spec == stored_resource.spec {
-            ResourceState::SpecNotChanged
-        } else {
-            ResourceState::SpecChanged
-        }
-    }
-
-    fn check_status(args: ResourceCheckerArgs<GRPCRoute>) -> ResourceState {
-        let (resource, stored_resource) = args;
-        if resource.status == stored_resource.status {
-            ResourceState::StatusNotChanged
-        } else {
-            ResourceState::StatusChanged
-        }
-    }
-}
-
-#[derive(TypedBuilder)]
-struct GRPCRouteHandler<R> {
-    state: State,
-    resource_key: ResourceKey,
-    controller_name: String,
-    resource: Arc<R>,
-    grpc_route_patcher: mpsc::Sender<Operation<GRPCRoute>>,
-    validate_references_channel_sender: mpsc::Sender<ReferenceValidateRequest>,
-    version: Option<String>,
-}
-
-#[async_trait]
-impl ResourceHandler<GRPCRoute> for GRPCRouteHandler<GRPCRoute> {
-    fn state(&self) -> &State {
-        &self.state
-    }
-
-    fn version(&self) -> String {
-        self.version.clone().unwrap_or_default()
-    }
-
-    fn resource(&self) -> Arc<GRPCRoute> {
-        Arc::clone(&self.resource)
-    }
-
-    fn resource_key(&self) -> ResourceKey {
-        self.resource_key.clone()
-    }
-
-    async fn on_spec_not_changed(&self, id: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        let () = state.save_grpc_route(id, resource).expect("We expect the lock to work");
-        Err(ControllerError::AlreadyAdded)
-    }
-
-    async fn on_status_not_changed(&self, id: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        let () = state.maybe_save_grpc_route(id, resource).expect("We expect the lock to work");
-        Err(ControllerError::AlreadyAdded)
-    }
-
-    async fn on_new(&self, id: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        self.on_new_or_changed(id, resource, state).await
-    }
-
-    async fn on_status_changed(&self, _id: ResourceKey, _resource: &Arc<GRPCRoute>, _state: &State) -> Result<Action> {
-        Ok(Action::await_change())
-    }
-
-    async fn on_spec_changed(&self, id: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        self.on_new_or_changed(id, resource, state).await
-    }
-
-    async fn on_deleted(&self, id: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        self.on_deleted(id, resource, state).await
-    }
-}
-
-impl GRPCRouteHandler<GRPCRoute> {
-    async fn on_new_or_changed(&self, route_key: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        let Some(parent_gateway_refs) = resource.spec.parent_refs.as_ref() else {
-            return Err(ControllerError::InvalidPayload("Route with no parents".to_owned()));
-        };
-
-        let mut grpc_route = (**resource).clone();
-        let route = Route::try_from(&grpc_route)?;
-
-        let parent_gateway_refs_keys = parent_gateway_refs.iter().map(|parent_ref| (parent_ref, RouteRefKey::from((parent_ref, route_key.namespace.clone()))));
-
-        let parent_gateway_refs = parent_gateway_refs_keys
-            .clone()
-            .map(|(parent_ref, key)| (parent_ref, state.get_gateway(key.as_ref()).expect("We expect the lock to work")))
-            .map(|i| if i.1.is_some() { Ok(i) } else { Err(i) });
-
-        let (resolved_gateways, unknown_gateways) = VerifiyItems::verify(parent_gateway_refs);
-
-        parent_gateway_refs_keys.for_each(|(_ref, key)| state.attach_grpc_route_to_gateway(key.as_ref().clone(), route_key.clone()).expect("We expect the lock to work"));
-
-        let matching_gateways = RouteListenerMatcher::filter_grpc_matching_gateways(state, &resolved_gateways);
-        let unknown_gateway_status = self.generate_status_for_unknown_gateways(&unknown_gateways, resource.metadata.generation);
-
-        grpc_route.status = Some(RouteStatus { parents: unknown_gateway_status });
-        let () = state.save_grpc_route(route_key.clone(), &Arc::new(grpc_route)).expect("We expect the lock to work");
-
-        let _ = self.add_finalizer(resource).await?;
-
-        let references = routes_common::extract_references(&route);
-
-        let _ = self
-            .validate_references_channel_sender
-            .send(ReferenceValidateRequest::AddRoute { references, span: Span::current() })
-            .await;
-
-        for kube_gateway in matching_gateways {
-            let gateway_class_name = {
-                let gateway_class_name = &kube_gateway.spec.gateway_class_name;
-
-                if !self
-                    .state
-                    .get_gateway_classes()
-                    .expect("We expect the lock to work")
-                    .into_iter()
-                    .any(|gc| gc.metadata.name == Some(gateway_class_name.clone()))
-                {
-                    warn!(
-                        "reconcile_gateway: {} {:?} Unknown gateway class name {gateway_class_name}",
-                        &self.controller_name,
-                        kube_gateway.meta().name
-                    );
-                    return Err(ControllerError::UnknownGatewayClass(gateway_class_name.clone()));
-                }
-                gateway_class_name.clone()
-            };
-            let kube_gateway = (*kube_gateway).clone();
-            let _ = self
-                .validate_references_channel_sender
-                .send(ReferenceValidateRequest::AddGateway(Box::new(
-                    RequestContext::builder()
-                        .gateway(common::Gateway::try_from(&kube_gateway).expect("We expect the lock to work"))
-                        .kube_gateway(kube_gateway)
-                        .gateway_class_name(gateway_class_name)
-                        .span(Span::current())
-                        .build(),
-                )))
-                .await;
-        }
-
-        Ok(Action::await_change())
-    }
-
-    async fn on_deleted(&self, route_key: ResourceKey, resource: &Arc<GRPCRoute>, state: &State) -> Result<Action> {
-        let _ = Route::try_from(&**resource)?;
-
-        let Some(parent_gateway_refs) = resource.spec.parent_refs.as_ref() else {
-            return Err(ControllerError::InvalidPayload("Route with no parents".to_owned()));
-        };
-
-        let parent_gateway_refs_keys = parent_gateway_refs.iter().map(|parent_ref| (parent_ref, RouteRefKey::from((parent_ref, route_key.namespace.clone()))));
-
-        debug!("Parent keys = {parent_gateway_refs_keys:?}");
-        parent_gateway_refs_keys.for_each(|(_ref, gateway_key)| state.detach_grpc_route_from_gateway(gateway_key.as_ref(), &route_key).expect("We expect the lock to work"));
-
-        let Some(route) = state.delete_grpc_route(&route_key).expect("We expect the lock to work") else {
-            return Err(ControllerError::InvalidPayload("Route doesn't exist".to_owned()));
-        };
-        let route = Route::try_from(&*route)?;
-
-        let grpc_route = (**resource).clone();
-        let resource_key = route_key;
-        let _res = self
-            .grpc_route_patcher
-            .send(Operation::Delete(DeleteContext {
-                resource_key: resource_key.clone(),
-                resource: grpc_route,
-                controller_name: self.controller_name.clone(),
-                span: Span::current().clone(),
-            }))
-            .await;
-
-        let references = routes_common::extract_references(&route);
-
-        let _ = self
-            .validate_references_channel_sender
-            .send(ReferenceValidateRequest::DeleteRoute { references, span: Span::current() })
-            .await;
-
-        Ok(Action::await_change())
-    }
-
-    async fn add_finalizer(&self, resource: &Arc<GRPCRoute>) -> Result<Action> {
-        if let Some(finalizer) = routes_common::needs_finalizer::<GRPCRoute>(&self.resource_key, &self.controller_name, &resource.metadata) {
-            let _res = self.grpc_route_patcher.send(finalizer).await;
-        }
-
-        Ok(Action::requeue(RECONCILE_LONG_WAIT))
-    }
-
-    fn generate_status_for_unknown_gateways(&self, gateways: &[(&RouteRef, Option<Arc<Gateway>>)], generation: Option<i64>) -> Vec<ParentRouteStatus> {
-        routes_common::generate_status_for_unknown_gateways(&self.controller_name, gateways, generation)
-    }
-}
