+++ src/controllers/route/mod.rs
@@ -1,0 +1,3 @@
+pub mod grpc_route;
+pub mod http_route;
+mod routes_common;
