+++ src/controllers/utils/route_listener_matcher.rs
@@ -100,7 +100,7 @@ impl<'a> RouteListenerMatcher<'a> {
         (routes_and_listeners, route_resolution_status)
     }
 
-    fn filter_matching_gateways(state: &State, resolved_gateways: &[(&RouteRef, Option<Arc<gateways::Gateway>>)]) -> Vec<Arc<gateways::Gateway>> {
+    pub fn filter_matching_gateways(state: &State, resolved_gateways: &[(&RouteRef, Option<Arc<gateways::Gateway>>)]) -> Vec<Arc<gateways::Gateway>> {
         resolved_gateways
             .iter()
             .filter_map(|(parent_ref, maybe_gateway)| {
@@ -120,13 +120,13 @@ impl<'a> RouteListenerMatcher<'a> {
             .collect()
     }
 
-    pub fn filter_http_matching_gateways(state: &State, resolved_gateways: &[(&RouteRef, Option<Arc<gateways::Gateway>>)]) -> Vec<Arc<gateways::Gateway>> {
-        Self::filter_matching_gateways(state, resolved_gateways)
-    }
+    // pub fn filter_http_matching_gateways(state: &State, resolved_gateways: &[(&RouteRef, Option<Arc<gateways::Gateway>>)]) -> Vec<Arc<gateways::Gateway>> {
+    //     Self::filter_matching_gateways(state, resolved_gateways)
+    // }
 
-    pub fn filter_grpc_matching_gateways(state: &State, resolved_gateways: &[(&RouteRef, Option<Arc<gateways::Gateway>>)]) -> Vec<Arc<gateways::Gateway>> {
-        Self::filter_matching_gateways(state, resolved_gateways)
-    }
+    // pub fn filter_grpc_matching_gateways(state: &State, resolved_gateways: &[(&RouteRef, Option<Arc<gateways::Gateway>>)]) -> Vec<Arc<gateways::Gateway>> {
+    //     Self::filter_matching_gateways(state, resolved_gateways)
+    // }
 
     fn filter_listeners_by_hostnames(listeners: impl Iterator<Item = GatewayListeners> + 'a, route: &'a Route) -> impl Iterator<Item = GatewayListeners> + 'a {
         let route_hostnames = route.hostnames();
